refactor: integration test renaming into system test
These are system-level test.
