+++ test/system/cli/__fixture__.mjs
@@ -8,7 +8,7 @@ import {
   realpath as realpathAsync,
 } from "fs/promises";
 import YAML from "yaml";
-import { spawnAsync } from "../spawn.mjs";
+import { spawnAsync } from "../../spawn.mjs";
 
 const { cwd } = process;
 const { stringify: stringifyYAML } = YAML;
