+++ test/system/cli/event-apply-source.mjs
@@ -4,7 +4,7 @@ import {
 } from "fs/promises";
 import { strict as Assert } from "assert";
 import { join as joinPath } from "path";
-import { runAsync } from "../__fixture__.mjs";
+import { runAsync } from "./__fixture__.mjs";
 import { SourceMapGenerator } from "source-map";
 
 const {
