+++ test/system/cli/npx.mjs
@@ -2,8 +2,8 @@ import { readFile as readFileAsync } from "fs/promises";
 import { platform as getPlatform } from "os";
 import { strict as Assert } from "assert";
 import { join as joinPath } from "path";
-import { spawnAsync } from "../spawn.mjs";
-import { runAsync } from "../__fixture__.mjs";
+import { spawnAsync } from "../../spawn.mjs";
+import { runAsync } from "./__fixture__.mjs";
 
 const { deepEqual: assertDeepEqual, equal: assertEqual } = Assert;
